Revert "Is logged in"
