+++ client/src/Components/Navbar.css
@@ -52,12 +52,9 @@ a.nav-signup:hover {
     cursor: pointer;
 }
 
-
-.material-symbols-outlined {
-    color: white;
-    padding: 0.5rem;
-}
-
+.account:hover {
     cursor: pointer;
 }
-
+.material-symbols-outlined {
+    color: white;
+}
