+++ client/src/Components/Navbar.jsx
@@ -1,12 +1,9 @@
-import {React, useContext} from "react";
+import {React} from "react";
 import { Link } from "react-router-dom";
 import "./Navbar.css";
 import { useNavigate } from "react-router-dom";
-import UserContext from "../UserContext";
 
 function Navbar() {
-  const auth = useContext(UserContext)
-
   const reset = (e) => {
     setSearchTerm("");
   };
@@ -14,7 +11,7 @@ function Navbar() {
   const navigate = useNavigate();
 
   function login() {
-    navigate("/login")
+      navigate("/login")
   }
 
   function signup() {
@@ -30,21 +27,11 @@ function Navbar() {
           <Link to="/" onClick={reset}>
             <h6 className="nav-title">Podcast App</h6>
           </Link>
-          <div className="m-1 mr-2 user-control">              
-              {!auth.isLoggedIn && (
-                <>
-                  <a className="nav-login p-2" onClick={login}>Login</a>
-                  <a className="nav-signup px-2" onClick={signup}>Sign up</a>
-                </>
-              )}
-              {auth.isLoggedIn && (
-                <>
-                  <span className="material-symbols-outlined" onClick={account}>account_circle</span>
-                  <a className="nav-signup px-2" onClick={auth.logout}>Logout</a>
-                </>
-              )}
-              
-              
+          <div className="m-1 mr-2">
+              {/* if we prefer google material symbols we can switch it out with the flaticons */}
+              <span className="material-symbols-outlined">account_circle</span>
+              <a className="nav-login p-2" onClick={login}>Login</a>
+              <a className="nav-signup px-2" onClick={signup}>Sign up</a>
           </div>
       </div>
 )};
