+++ client/src/Pages/Account.jsx
@@ -2,7 +2,6 @@
 import Navbar from "../Components/Navbar";
 
 export default function Account() {
-    const auth = useContext(UserContext)
 
     return(
         <>
